Fix currency implementation

+++ BlockchainGateway/bc-gateway-ports/bc-persister-postgres/src/main/kotlin/co/nilin/opex/port/bcgateway/postgres/model/CurrencyImplementationModel.kt
@@ -4,6 +4,7 @@ package co.nilin.opex.port.bcgateway.postgres.model
 import org.springframework.data.annotation.Id
 import org.springframework.data.relational.core.mapping.Column
 import org.springframework.data.relational.core.mapping.Table
+import java.math.BigDecimal
 
 @Table("currency_implementations")
 class CurrencyImplementationModel(
@@ -14,6 +15,6 @@ class CurrencyImplementationModel(
     @Column("token_address") val tokenAddress: String?,
     @Column("token_name") val tokenName: String?,
     @Column("withdraw_enabled") val withdrawEnabled: Boolean,
-    @Column("withdraw_fee") val withdrawFee: Long,
-    @Column("withdraw_min") val withdrawMin: Long,
+    @Column("withdraw_fee") val withdrawFee: BigDecimal,
+    @Column("withdraw_min") val withdrawMin: BigDecimal,
 )
